// client/src/data/Interfaces/ISuccessPage.ts
export interface ISuccessPage {
  Title: string;
  SubTitle: string;
  Id: string;
  RentalName: string;
  Location: string;
  StartDate: string;
  EndDate: string;
  NumberOfNights: number;
  NumberOfPeople: number;
  PaymentMethod: string;
  HomeButton: string;
  OrderButton: string;
}